% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-if.R
\name{examplesIf}
\alias{examplesIf}
\title{Conditional examples}
\description{
Conditional examples

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
The \verb{@examplesIf} tag implements conditional examples, that do not run
when running \code{example()} and \verb{R CMD check}, unless some condition holds.
You can use this mechanism to avoid running some examples on CRAN and/or
during \verb{R CMD check}, without adding boilerplate code to the examples
themselves. The following example only runs if the computer is online:\preformatted{#' @examplesIf curl::has_internet()
#' curl::curl_fetch_memory("https://httpbin.org/status/200")
}

If you never want to run an example, use \code{FALSE} as the condition. This
makes sense for examples that take a very long time to run:\preformatted{#' @examplesIf FALSE
#' some very long running computation...
}

To support \verb{@examplesIf} in R6 classes you need to use the
\code{roxygenlabs_rd} roclet. E.g. add this to \code{DESCRIPTION}:\preformatted{Roxygen: \{ library(roxygenlabs); list(markdown = TRUE,
    roclets = c("collate", "namespace", "roxygenlabs_rd")) \}
}
}
