% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{roxygenlabs-package}
\alias{roxygenlabs}
\alias{roxygenlabs-package}
\title{Experimental 'roxygen' tags and extensions}
\description{
Experimental 'roxygen' tags and extensions.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/roxygenlabs}
  \item Report bugs at \url{https://github.com/gaborcsardi/roxygenlabs/issues}
}

}
