% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roxygenlabs_rd}
\alias{roxygenlabs_rd}
\title{Use this roclet for roxygenlabs tags and extensions}
\usage{
roxygenlabs_rd()
}
\description{
To use this \code{roxygenlabs_rd} roclet, add this to \code{DESCRIPTION}:

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
\preformatted{Roxygen: \{ library(roxygenlabs); list(markdown = TRUE,
    roclets = c("collate", "namespace", "roxygenlabs_rd")) \}
}
}
