% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themed-rd.R
\name{themed_rd}
\alias{themed_rd}
\title{Theme R manual pages with roxygen2}
\usage{
themed_rd(test)
}
\arguments{
\item{test}{This is a test.}
}
\value{
Roxygen roclet.
}
\description{
This roclet allows adding CSS and JS files to the HTML version of the
manual pages.

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\details{
To use this roclet, you need to add this to \code{DESCRIPTION}:\preformatted{Roxygen: \{ library(roxygenlabs); list(markdown = TRUE,
    roclets = c("collate", "namespace", "themed_rd")) \}
}

and then use the \verb{@theme} tag to add CSS/JS files to \emph{all} manual pages.
Multiple file names can be specicified in a single \verb{@theme} tag, and
file names are relative to the \code{inst/doc} directory, within the package.
}
