% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{roxygenlabs-package}
\alias{roxygenlabs}
\alias{roxygenlabs-package}
\title{Experimental 'roxygen' Tags and Extensions}
\description{
Experimental 'roxygen' tags and extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/roxygenlabs}
  \item Report bugs at \url{https://github.com/gaborcsardi/roxygenlabs/issues}
}

}
